django_api_parafernalia/requirements.txt
asgiref>=3.2.10
certifi>=2020.6.20
chardet>=3.0.4
Django>=3.0.8
django-model-utils>=4.0.0
djangorestframework>=3.11.0
idna>=2.10
psycopg2>=2.8.5
pytz>=2020.1
requests>=2.24.0
sqlparse>=0.3.1
urllib3>=1.25.9